app/requirements.txt
fastapi
uvicorn
psycopg2-binary
# psycopg2

